fix missing javadoc and relocate the common compress to compress package
