+++ gradle/missing-javadoc.gradle
@@ -166,6 +166,7 @@ configure([
 configure([
   project(":libs:opensearch-common"),
   project(":libs:opensearch-core"),
+  project(":libs:opensearch-compress"),
   project(":plugins:events-correlation-engine"),
   project(":server")
 ]) {
