+++ libs/compress/src/main/java/org/opensearch/compress/spi/package-info.java
@@ -1,0 +1,15 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * Service Provider Interface for registering concrete {@link org.opensearch.core.compress.Compressor}
+ * implementations.
+ *
+ * See {@link org.opensearch.compress.ZstdCompressor}
+ */
+package org.opensearch.compress.spi;
