+++ libs/compress/src/main/java/org/opensearch/package-info.java
@@ -6,5 +6,8 @@
  * compatible open source license.
  */
 
-/** Classes for core compress module */
-package org.opensearch.core.common.compress;
+/**
+ * This is the compress library for registering optional
+ * {@link org.opensearch.core.compress.Compressor} implementations
+ */
+package org.opensearch;
