+++ libs/core/src/main/java/org/opensearch/core/compress/NoneCompressor.java
@@ -9,7 +9,6 @@
 package org.opensearch.core.compress;
 
 import org.opensearch.core.common.bytes.BytesReference;
-import org.opensearch.core.common.compress.Compressor;
 
 import java.io.IOException;
 import java.io.InputStream;
