+++ libs/core/src/main/java/org/opensearch/core/compress/spi/CompressorProvider.java
@@ -8,7 +8,7 @@
 
 package org.opensearch.core.compress.spi;
 
-import org.opensearch.core.common.compress.Compressor;
+import org.opensearch.core.compress.Compressor;
 import org.opensearch.core.compress.CompressorRegistry;
 
 import java.util.List;
