+++ libs/core/src/main/java/org/opensearch/core/compress/spi/DefaultCompressorProvider.java
@@ -8,19 +8,20 @@
 
 package org.opensearch.core.compress.spi;
 
+import org.opensearch.core.compress.Compressor;
 import org.opensearch.core.compress.CompressorRegistry;
 import org.opensearch.core.compress.NoneCompressor;
 
 import java.util.List;
 
 /**
- * Default {@link org.opensearch.core.common.compress.Compressor} implementations provided by the
+ * Default {@link Compressor} implementations provided by the
  * opensearch core library
  *
  * @opensearch.internal
  */
 public class DefaultCompressorProvider implements CompressorProvider {
-    /** Returns the default {@link org.opensearch.core.common.compress.Compressor}s provided by the core library */
+    /** Returns the default {@link Compressor}s provided by the core library */
     @Override
     public List<CompressorRegistry.Entry> getCompressors() {
         return List.of(new CompressorRegistry.Entry(NoneCompressor.NAME, new NoneCompressor()));
