+++ libs/core/src/main/java/org/opensearch/core/compress/spi/package-info.java
@@ -1,0 +1,16 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * The Service Provider Interface implementation for registering {@link org.opensearch.core.compress.Compressor}
+ * with the {@link org.opensearch.core.compress.CompressorRegistry}
+ *
+ * See {@link org.opensearch.core.compress.spi.DefaultCompressorProvider} as an example of registering the core
+ * {@link org.opensearch.core.compress.NoneCompressor}
+ */
+package org.opensearch.core.compress.spi;
