+++ server/src/main/java/org/opensearch/cluster/coordination/CompressedStreamUtils.java
@@ -14,7 +14,7 @@ import org.opensearch.Version;
 import org.opensearch.common.CheckedConsumer;
 import org.opensearch.common.io.stream.BytesStreamOutput;
 import org.opensearch.core.common.bytes.BytesReference;
-import org.opensearch.core.common.compress.Compressor;
+import org.opensearch.core.compress.Compressor;
 import org.opensearch.core.common.io.stream.InputStreamStreamInput;
 import org.opensearch.core.common.io.stream.NamedWriteableAwareStreamInput;
 import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
