+++ server/src/main/java/org/opensearch/common/compress/spi/ServerCompressorProvider.java
@@ -9,13 +9,14 @@
 package org.opensearch.common.compress.spi;
 
 import org.opensearch.common.compress.DeflateCompressor;
+import org.opensearch.core.compress.Compressor;
 import org.opensearch.core.compress.CompressorRegistry;
 import org.opensearch.core.compress.spi.CompressorProvider;
 
 import java.util.List;
 
 /**
- * Default {@link org.opensearch.core.common.compress.Compressor} implementations provided by the
+ * Default {@link Compressor} implementations provided by the
  * opensearch core library
  *
  * @opensearch.internal
@@ -25,7 +26,7 @@ import java.util.List;
  */
 @Deprecated
 public class ServerCompressorProvider implements CompressorProvider {
-    /** Returns the concrete {@link org.opensearch.core.common.compress.Compressor}s provided by the server module */
+    /** Returns the concrete {@link Compressor}s provided by the server module */
     @Override
     public List<CompressorRegistry.Entry> getCompressors() {
         return List.of(new CompressorRegistry.Entry(DeflateCompressor.NAME, new DeflateCompressor()));
