+++ server/src/main/java/org/opensearch/repositories/blobstore/BlobStoreRepository.java
@@ -80,8 +80,8 @@ import org.opensearch.core.common.bytes.BytesArray;
 import org.opensearch.core.common.bytes.BytesReference;
 import org.opensearch.common.collect.Tuple;
 import org.opensearch.common.lifecycle.AbstractLifecycleComponent;
-import org.opensearch.core.common.compress.Compressor;
-import org.opensearch.core.common.compress.NotXContentException;
+import org.opensearch.core.compress.Compressor;
+import org.opensearch.core.compress.NotXContentException;
 import org.opensearch.common.io.Streams;
 import org.opensearch.common.lucene.Lucene;
 import org.opensearch.common.lucene.store.InputStreamIndexInput;
