+++ server/src/test/java/org/opensearch/common/compress/DeflateCompressTests.java
@@ -32,7 +32,7 @@
 
 package org.opensearch.common.compress;
 
-import org.opensearch.core.common.compress.Compressor;
+import org.opensearch.core.compress.Compressor;
 import org.opensearch.test.core.compress.AbstractCompressorTestCase;
 
 /**
