+++ server/src/test/java/org/opensearch/common/compress/DeflateCompressedXContentTests.java
@@ -35,7 +35,7 @@ package org.opensearch.common.compress;
 import org.apache.lucene.tests.util.TestUtil;
 import org.opensearch.core.common.bytes.BytesReference;
 import org.opensearch.common.io.stream.BytesStreamOutput;
-import org.opensearch.core.common.compress.Compressor;
+import org.opensearch.core.compress.Compressor;
 import org.opensearch.test.OpenSearchTestCase;
 import org.junit.Assert;
